requirements.txt
torch
surya-ocr
pillow
flask==2.0.1
reportlab==4.0.0
werkzeug==2.0.1
requests==2.28.1
uuid==1.30